Revert only under the `test` dir: "refactor: use import syntax"
This reverts commit b1a4c57b9b6fea26fc839e94b7f1d5ef94697291.
