+++ test/scripts/document.js
@@ -1,9 +1,10 @@
-import chai from 'chai';
-const should = chai.should(); // eslint-disable-line
-import Database from '../../lib/database';
-import Document from '../../lib/document';
+'use strict';
+
+const should = require('chai').should(); // eslint-disable-line
 
 describe('Document', () => {
+  const Database = require('../..');
+  const Document = require('../../built/document');
   const db = new Database();
   const Schema = Database.Schema;
