+++ test/scripts/schematype.js
@@ -1,9 +1,10 @@
-import chai from 'chai';
-const should = chai.should(); // eslint-disable-line
-import ValidationError from '../../lib/error/validation';
-import SchemaType from '../../lib/schematype';
+'use strict';
+
+const should = require('chai').should(); // eslint-disable-line
+const ValidationError = require('../../built/error/validation');
 
 describe('SchemaType', () => {
+  const SchemaType = require('../../built/schematype');
   const type = new SchemaType('test');
 
   it('cast()', () => {
