+++ test/scripts/types/boolean.js
@@ -1,9 +1,10 @@
-import chai from 'chai';
-const should = chai.should(); // eslint-disable-line
-import ValidationError from '../../../lib/error/validation';
-import SchemaTypeBoolean from '../../../lib/types/boolean';
+'use strict';
+
+require('chai').should();
+const ValidationError = require('../../../built/error/validation');
 
 describe('SchemaTypeBoolean', () => {
+  const SchemaTypeBoolean = require('../../../built/types/boolean');
   const type = new SchemaTypeBoolean('test');
 
   it('cast()', () => {
