+++ test/scripts/types/integer.js
@@ -1,9 +1,10 @@
-import chai from 'chai';
-const should = chai.should(); // eslint-disable-line
-import ValidationError from '../../../lib/error/validation';
-import SchemaTypeInteger from '../../../lib/types/integer';
+'use strict';
+
+require('chai').should();
+const ValidationError = require('../../../built/error/validation');
 
 describe('SchemaTypeInteger', () => {
+  const SchemaTypeInteger = require('../../../built/types/integer');
   const type = new SchemaTypeInteger('test');
 
   it('cast()', () => {
