+++ test/scripts/types/string.js
@@ -1,9 +1,10 @@
-import chai from 'chai';
-const should = chai.should(); // eslint-disable-line
-import ValidationError from '../../../lib/error/validation';
-import SchemaTypeString from '../../../lib/types/string';
+'use strict';
+
+require('chai').should();
+const ValidationError = require('../../../built/error/validation');
 
 describe('SchemaTypeString', () => {
+  const SchemaTypeString = require('../../../built/types/string');
   const type = new SchemaTypeString('test');
 
   it('cast()', () => {
